+++ packages/db/test/db.spec.ts
@@ -32,12 +32,6 @@ import {
   appendCollectionToId,
   stripCollectionFromId,
 } from '../src/db-helpers.js';
-import {
-  CollectionNameFromQuery,
-  JSTypeOrRelation,
-} from '../src/collection-query.js';
-import { CollectionFromModels, ModelFromModels } from '../src/db.js';
-import { set } from 'tuple-database/helpers/sortedTupleArray.js';
 
 const pause = async (ms: number = 100) =>
   new Promise((resolve) => setTimeout(resolve, ms));
@@ -5282,7 +5276,7 @@ describe('selecting subqueries from schema', () => {
             id: S.String(),
             content: S.String(),
             author_id: S.String(),
-            author: S.Entity('users', '$author_id'),
+            author: S.RelationById('users', '$author_id'),
             topics: S.Set(S.String()),
             likes: S.Query({
               collectionName: 'likes' as const,
@@ -5514,5 +5508,231 @@ describe('selecting subqueries from schema', () => {
         },
       },
     ]);
+  });
+});
+
+describe('hooks API', async () => {
+  it('before write hooks will run on transaction', async () => {
+    const db = new DB();
+    const beforeCommitFn = vi.fn();
+    db.addTrigger(
+      { when: 'beforeCommit', collectionName: 'users' },
+      beforeCommitFn
+    );
+    const beforeInsertFn = vi.fn();
+    db.addTrigger(
+      { when: 'beforeInsert', collectionName: 'users' },
+      beforeInsertFn
+    );
+    const beforeUpdateFn = vi.fn();
+    db.addTrigger(
+      { when: 'beforeUpdate', collectionName: 'users' },
+      beforeUpdateFn
+    );
+    const beforeDeleteFn = vi.fn();
+    db.addTrigger(
+      { when: 'beforeDelete', collectionName: 'users' },
+      beforeDeleteFn
+    );
+    await db.transact(async (tx) => {
+      await tx.insert('users', { id: '1', name: 'alice' });
+      await tx.insert('users', { id: '2', name: 'bob' });
+    });
+    expect(beforeCommitFn).toHaveBeenCalledTimes(1);
+    expect(beforeCommitFn.mock.calls[0][0].opSet).toStrictEqual({
+      inserts: [
+        ['users#1', { _collection: 'users', id: '1', name: 'alice' }],
+        ['users#2', { _collection: 'users', id: '2', name: 'bob' }],
+      ],
+      updates: [],
+      deletes: [],
+    });
+    expect(beforeInsertFn).toHaveBeenCalledTimes(2);
+    expect(beforeInsertFn.mock.calls[0][0].entity).toStrictEqual({
+      _collection: 'users',
+      id: '1',
+      name: 'alice',
+    });
+    expect(beforeInsertFn.mock.calls[1][0].entity).toStrictEqual({
+      _collection: 'users',
+      id: '2',
+      name: 'bob',
+    });
+    expect(beforeUpdateFn).toHaveBeenCalledTimes(0);
+    expect(beforeDeleteFn).toHaveBeenCalledTimes(0);
+    await db.transact(async (tx) => {
+      await tx.update('users', '1', (entity) => {
+        entity.name = 'aaron';
+      });
+      await tx.update('users', '2', (entity) => {
+        entity.name = 'blair';
+      });
+    });
+    expect(beforeCommitFn).toHaveBeenCalledTimes(2);
+    expect(beforeCommitFn.mock.calls[1][0].opSet).toStrictEqual({
+      inserts: [],
+      updates: [
+        ['users#1', { _collection: 'users', id: '1', name: 'aaron' }],
+        ['users#2', { _collection: 'users', id: '2', name: 'blair' }],
+      ],
+      deletes: [],
+    });
+    expect(beforeInsertFn).toHaveBeenCalledTimes(2);
+    expect(beforeUpdateFn).toHaveBeenCalledTimes(2);
+    expect(beforeUpdateFn.mock.calls[0][0].entity).toStrictEqual({
+      _collection: 'users',
+      id: '1',
+      name: 'aaron',
+    });
+    expect(beforeUpdateFn.mock.calls[1][0].entity).toStrictEqual({
+      _collection: 'users',
+      id: '2',
+      name: 'blair',
+    });
+    expect(beforeDeleteFn).toHaveBeenCalledTimes(0);
+    await db.transact(async (tx) => {
+      await tx.delete('users', '1');
+      await tx.delete('users', '2');
+    });
+    expect(beforeCommitFn).toHaveBeenCalledTimes(3);
+    expect(beforeCommitFn.mock.calls[2][0].opSet).toStrictEqual({
+      inserts: [],
+      updates: [],
+      deletes: [
+        ['users#1', { id: '1', name: 'aaron' }],
+        ['users#2', { id: '2', name: 'blair' }],
+      ],
+    });
+    expect(beforeInsertFn).toHaveBeenCalledTimes(2);
+    expect(beforeUpdateFn).toHaveBeenCalledTimes(2);
+    expect(beforeDeleteFn).toHaveBeenCalledTimes(2);
+    expect(beforeDeleteFn.mock.calls[0][0].entity).toStrictEqual({
+      id: '1',
+      name: 'aaron',
+    });
+    expect(beforeDeleteFn.mock.calls[1][0].entity).toStrictEqual({
+      id: '2',
+      name: 'blair',
+    });
+  });
+  it('after write hooks will run on transaction', async () => {
+    const db = new DB({
+      schema: {
+        collections: {
+          users: {
+            schema: S.Schema({
+              id: S.String(),
+              name: S.String(),
+            }),
+          },
+          tasks: {
+            schema: S.Schema({
+              id: S.String(),
+              text: S.String(),
+              due: S.Date(),
+              completed: S.Boolean(),
+            }),
+          },
+        },
+      },
+    });
+    const afterCommitFn = vi.fn();
+    db.addTrigger({ when: 'afterCommit' }, afterCommitFn);
+    const afterInsertFn = vi.fn();
+    db.addTrigger(
+      { when: 'afterInsert', collectionName: 'users' },
+      afterInsertFn
+    );
+    const afterUpdateFn = vi.fn();
+    db.addTrigger(
+      { when: 'afterUpdate', collectionName: 'users' },
+      afterUpdateFn
+    );
+    const afterDeleteFn = vi.fn();
+    db.addTrigger(
+      { when: 'afterDelete', collectionName: 'users' },
+      afterDeleteFn
+    );
+
+    await db.transact(async (tx) => {
+      await tx.insert('users', { id: '1', name: 'alice' });
+      await tx.insert('users', { id: '2', name: 'bob' });
+    });
+    expect(afterCommitFn).toHaveBeenCalledTimes(1);
+    expect(afterCommitFn.mock.calls[0][0].opSet).toStrictEqual({
+      inserts: [
+        ['users#1', { _collection: 'users', id: '1', name: 'alice' }],
+        ['users#2', { _collection: 'users', id: '2', name: 'bob' }],
+      ],
+      updates: [],
+      deletes: [],
+    });
+    expect(afterInsertFn).toHaveBeenCalledTimes(2);
+    expect(afterInsertFn.mock.calls[0][0].entity).toStrictEqual({
+      _collection: 'users',
+      id: '1',
+      name: 'alice',
+    });
+    expect(afterInsertFn.mock.calls[1][0].entity).toStrictEqual({
+      _collection: 'users',
+      id: '2',
+      name: 'bob',
+    });
+    expect(afterUpdateFn).toHaveBeenCalledTimes(0);
+    expect(afterDeleteFn).toHaveBeenCalledTimes(0);
+    await db.transact(async (tx) => {
+      await tx.update('users', '1', (entity) => {
+        entity.name = 'aaron';
+      });
+      await tx.update('users', '2', (entity) => {
+        entity.name = 'blair';
+      });
+    });
+    expect(afterCommitFn).toHaveBeenCalledTimes(2);
+    expect(afterCommitFn.mock.calls[1][0].opSet).toStrictEqual({
+      inserts: [],
+      updates: [
+        ['users#1', { _collection: 'users', id: '1', name: 'aaron' }],
+        ['users#2', { _collection: 'users', id: '2', name: 'blair' }],
+      ],
+      deletes: [],
+    });
+    expect(afterInsertFn).toHaveBeenCalledTimes(2);
+    expect(afterUpdateFn).toHaveBeenCalledTimes(2);
+    expect(afterUpdateFn.mock.calls[0][0].entity).toStrictEqual({
+      _collection: 'users',
+      id: '1',
+      name: 'aaron',
+    });
+    expect(afterUpdateFn.mock.calls[1][0].entity).toStrictEqual({
+      _collection: 'users',
+      id: '2',
+      name: 'blair',
+    });
+    expect(afterDeleteFn).toHaveBeenCalledTimes(0);
+    await db.transact(async (tx) => {
+      await tx.delete('users', '1');
+      await tx.delete('users', '2');
+    });
+    expect(afterCommitFn).toHaveBeenCalledTimes(3);
+    expect(afterCommitFn.mock.calls[2][0].opSet).toStrictEqual({
+      inserts: [],
+      updates: [],
+      deletes: [
+        ['users#1', { id: '1', name: 'aaron' }],
+        ['users#2', { id: '2', name: 'blair' }],
+      ],
+    });
+    expect(afterInsertFn).toHaveBeenCalledTimes(2);
+    expect(afterUpdateFn).toHaveBeenCalledTimes(2);
+    expect(afterDeleteFn).toHaveBeenCalledTimes(2);
+    expect(afterDeleteFn.mock.calls[0][0].entity).toStrictEqual({
+      id: '1',
+      name: 'aaron',
+    });
+    expect(afterDeleteFn.mock.calls[1][0].entity).toStrictEqual({
+      id: '2',
+      name: 'blair',
+    });
   });
 });