+++ templates/chat-template/components/client-auth-provider.tsx
@@ -13,7 +13,9 @@ export function ClientAuthProvider({
   const { data: session } = useSession()
   const jwtRef = useRef<string | undefined>()
   useEffect(() => {
+    // @ts-ignore
     if (session?.token !== jwtRef.current) {
+      // @ts-ignore
       jwtRef.current = session?.token ?? undefined
       client.updateToken(jwtRef.current)
     }