+++ packages/server-core/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @triplit/server-core
 
+## 0.3.11
+
+### Patch Changes
+
+- Updated dependencies [ae9bad9]
+  - @triplit/db@0.3.10
+
+## 0.3.10
+
+### Patch Changes
+
+- 7c8e6eb: add option to use external jwt secret for token verification
+
+## 0.3.9
+
+### Patch Changes
+
+- Updated dependencies [ff3bfe2]
+  - @triplit/db@0.3.9
+
+## 0.3.8
+
+### Patch Changes
+
+- Updated dependencies [f4f87df]
+  - @triplit/db@0.3.8
+
+## 0.3.7
+
+### Patch Changes
+
+- Updated dependencies [4d2d381]
+  - @triplit/db@0.3.7
+
 ## 0.3.6
 
 ### Patch Changes